+++ QueenCam/QueenCam/Presentation/ViewModels/WifiAware/ConnectionViewModel.swift
@@ -31,7 +31,7 @@ final class ConnectionViewModel {
     didSet {
       // 연결 상태 변화에 따른 State Toast 처리
       // 그 외 사이드 이펙트가 따라오는 다른 작업은 최대한 피할 것
-      
+
       if networkState == .host(.stopped) || networkState == .viewer(.stopped) {
         notificationService.registerBaseNotification(DomainNotification.make(type: .ready))
       } else {
@@ -52,6 +52,10 @@ final class ConnectionViewModel {
 
   /// 연결 유실 여부를 표현하는 플래그. true이면 재연결을 시작하고 관련 UI를 표시
   var connectionLost: Bool = false
+  /// 재연결 중인 디바이스 이름
+  var reconnectingDeviceName: String?
+  
+  var needReportSessionFinished: Bool = false
 
   /// 최근 역할 스왑 LWW 기록
   private var lastSwapRoleLWWRegister: LWWRegister?
@@ -92,6 +96,7 @@ final class ConnectionViewModel {
         // 이벤트 전파 후 핸들링
         if state == .host(.lost) || state == .viewer(.lost) {
           connectionLost = true
+          reconnectingDeviceName = lastConnectedDevice?.name
           tryReconnect()
         }
       }
@@ -116,6 +121,9 @@ final class ConnectionViewModel {
           self?.lastPingAt = pingAt
         case .changeRole(let roles, let lwwValue):
           self?.handleReceivedRequestChangeRole(receivedNewRoles: roles, receviedLwwRegister: lwwValue)
+        case .willDisconnect:
+          // 상대로부터 연결 종료 예정 통보를 받으면 세션 종료 오버레이 노출
+          self?.needReportSessionFinished = true
         default: break
         }
       }
@@ -200,6 +208,11 @@ extension ConnectionViewModel {
     networkService.stop(byUser: true)
     lastConnectedDevice = nil
     connectionLost = false
+    reconnectingDeviceName = nil
+  }
+  
+  func sessionFinishedOverlayCloseButtonDidTap() {
+    needReportSessionFinished = false
   }
 }
 
@@ -209,6 +222,7 @@ extension ConnectionViewModel {
     if let firstConnection = connections.first {
       lastConnectedDevice = firstConnection.key
       connectionLost = false  // 재연결인 경우 connectionLost 플래그를 초기화
+      reconnectingDeviceName = nil
     }
   }