+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 - 펜 수정 툴바(전체삭제, 실행 취소) 노출 및 활성화 조건 추가
 - 매직펜 하이파이 디자인 추가(블러효과)
 - 권한 거부에 따른 UI 추가
+- 하드웨어 버튼을 통한 촬영
+- 프레임 비율 조정 시, 프레임 내 그리드와 디밍 효과 추가
+- 양방향 프레임의 수정 점유권에 따른 하이파이 구현
+- 정상적으로 상대가 연결을 종료했을 때 오버레이 표시
 
 ### Changed
 
@@ -24,12 +28,19 @@
 - 펜과 매직펜 구분 프로퍼티 추가
 - 화면 내 레퍼런스 위치 및 배치 수정
 - 원본 이미지의 비율을 고려하여 레퍼런스 크기 조절
+- 재연결 뷰 하이파이 디자인 적용
+- 손가락 위치를 기준으로 핀치 제스처 작동
+- Target에서 iPad, Mac 제외
+- Orientation을 Portrait로 고정
 
 ### Deprecated
 
 
 ### Removed
 
+- FrameDisplayView
+- Frame의 Color 프로퍼티
+- NetworkStateModalView 제거하고, 시스템 기본 UI인 Menu로 대체
 
 ### Fixed
 
@@ -37,6 +48,8 @@
 - PhotosPickerView에서 썸네일 이미지를 레퍼런스 이미지로 선택하면 정방형 이미지가 로드되던 문제
 - PhotosPickerView와 PhotoDetailView 체크박스 상태 동기화 이슈 해결
 - 레퍼런스 삭제시 PhotoPicker 초기화가 안된 문제 해결
+-  프레임의 동시 수정 관련된 문제 해결
+- SampleBufferDisplayView에서 매번 화질 관련 지표를 리셋하여 적응형 화질 매커니즘이 동작하지 않았던 문제
 
 ### Security